Don't transform when a property name is an integer
An error (`TypeError: Value needs to be a string`) was raised when a
property name was an integer (`obj[10]`).

I don't know if it's the best way to check if we can transform or not a
property. Currently I check if `node.property.value` is not a number.
Should I check `startToken.type === 'Numeric'` instead?